Move AbstractBundle to ExtensionsModule
