+++ src/Zikula/CoreBundle/HttpKernel/ZikulaKernel.php
@@ -22,8 +22,8 @@ use Zikula\AdminModule\ZikulaAdminModule;
 use Zikula\AtomTheme\ZikulaAtomTheme;
 use Zikula\BlocksModule\ZikulaBlocksModule;
 use Zikula\BootstrapTheme\ZikulaBootstrapTheme;
-use Zikula\Bundle\CoreBundle\AbstractBundle;
 use Zikula\CategoriesModule\ZikulaCategoriesModule;
+use Zikula\ExtensionsModule\AbstractBundle;
 use Zikula\ExtensionsModule\AbstractModule;
 use Zikula\ExtensionsModule\ZikulaExtensionsModule;
 use Zikula\GroupsModule\ZikulaGroupsModule;
