+++ src/system/ExtensionsModule/Installer/AbstractExtensionInstaller.php
@@ -17,9 +17,9 @@ use Doctrine\ORM\EntityManagerInterface;
 use LogicException;
 use Symfony\Component\DependencyInjection\ContainerAwareInterface;
 use Symfony\Component\DependencyInjection\ContainerInterface;
-use Zikula\Bundle\CoreBundle\AbstractBundle;
 use Zikula\Bundle\CoreBundle\Doctrine\Helper\SchemaHelper;
 use Zikula\Bundle\CoreBundle\Translation\TranslatorTrait;
+use Zikula\ExtensionsModule\AbstractBundle;
 use Zikula\ExtensionsModule\Api\VariableApi;
 use Zikula\ExtensionsModule\ExtensionVariablesTrait;
 
@@ -42,7 +42,7 @@ abstract class AbstractExtensionInstaller implements ExtensionInstallerInterface
     protected $container;
 
     /**
-     * @var AbstractBundle
+     * @var \Zikula\ExtensionsModule\AbstractBundle
      */
     protected $bundle;
