+++ src/system/ThemeModule/AbstractTheme.php
@@ -17,7 +17,7 @@ use Symfony\Component\Config\Definition\Exception\InvalidConfigurationException;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Yaml\Yaml;
 use Twig\Environment;
-use Zikula\Bundle\CoreBundle\AbstractBundle;
+use Zikula\ExtensionsModule\AbstractBundle;
 use Zikula\ExtensionsModule\Api\VariableApi;
 
 abstract class AbstractTheme extends AbstractBundle
